:recycle: refactor: 컴포넌트의 interface refactoring
