+++ src/components/button/index.tsx
@@ -1,5 +1,9 @@
-import { IButton } from "src/interfaces/component";
 import { Container } from "./index.style";
+
+export interface IButton {
+  selected: boolean;
+  children: string;
+}
 
 function Button({ selected, children }: IButton) {
   return <Container selected={selected}>{children}</Container>;
