+++ src/components/loading/index.tsx
@@ -1,5 +1,11 @@
-import { ILoading } from "src/interfaces/component";
+import { ReactNode } from "react";
 import { Container, IconContainer } from "./index.style";
+
+export interface ILoading {
+  title: ReactNode;
+  icon: ReactNode;
+  bottomInformation?: ReactNode;
+}
 
 const Loading = ({ title, icon, bottomInformation }: ILoading) => {
   return (
