+++ src/components/tag/index.tsx
@@ -1,5 +1,9 @@
-import { ITag } from "src/interfaces/component";
 import { Container } from "./index.style";
+
+export interface ITag {
+  children: string;
+  selected: boolean;
+}
 
 const Tag = ({ children, selected }: ITag) => {
   return <Container selected={selected}>{children}</Container>;
