+++ src/components/treatmentTag/index.tsx
@@ -1,5 +1,9 @@
 import { Container } from "./index.style";
-import { ITreatmentTag } from "src/interfaces/component";
+
+export type ITreatmentType = "therapy" | "inspection";
+export interface ITreatmentTag {
+  type: ITreatmentType;
+}
 
 const TreatmentTag = ({ type }: ITreatmentTag) => {
   return <Container type={type}>{type === "therapy" ? "치료" : "검사"}</Container>;
